requirements.txt
torchvision
numpy
pandas
scikit-learn
matplotlib
pillow
kagglehub
tqdm